airflow/requirements.txt
requests
kafka-python
xmltodict
apache-airflow-providers-docker